init refactor multi language controlling to language_support, and pre_commit only manages one langauge
