+++ secureli/actions/action.py
@@ -18,7 +18,6 @@ from secureli.services.language_analyzer import LanguageAnalyzerService, Analyze
 from secureli.services.language_support import LanguageSupportService
 from secureli.services.scanner import ScannerService, ScanMode
 from secureli.services.updater import UpdaterService
-from secureli.abstractions.pre_commit import PreCommitAbstraction
 
 
 class VerifyOutcome(str, Enum):
@@ -60,7 +59,6 @@ class ActionDependencies:
         scanner: ScannerService,
         secureli_config: SecureliConfigRepository,
         updater: UpdaterService,
-        pre_commit: PreCommitAbstraction,
     ):
         self.echo = echo
         self.language_analyzer = language_analyzer
@@ -68,7 +66,6 @@ class ActionDependencies:
         self.scanner = scanner
         self.secureli_config = secureli_config
         self.updater = updater
-        self.pre_commit = pre_commit
 
 
 class Action(ABC):
@@ -89,11 +86,11 @@ class Action(ABC):
 
         config = SecureliConfig() if reset else self.action_deps.secureli_config.load()
 
-        if not config.overall_language or not config.version_installed:
+        if not config.languages or not config.version_installed:
             return self._install_secureli(folder_path, always_yes)
         else:
             available_version = self.action_deps.language_support.version_for_language(
-                config.overall_language
+                config.languages
             )
 
             # Check for a new version and prompt for upgrade if available
@@ -101,8 +98,10 @@ class Action(ABC):
                 return self._upgrade_secureli(config, available_version, always_yes)
 
             # Validates the current .pre-commit-config.yaml against the generated config
-            config_validation_result = self.action_deps.pre_commit.validate_config(
-                language=config.overall_language
+            config_validation_result = (
+                self.action_deps.language_support.validate_config(
+                    language=config.languages
+                )
             )
 
             # If config mismatch between available version and current version prompt for upgrade
@@ -111,7 +110,7 @@ class Action(ABC):
                 return self._update_secureli(always_yes)
 
             self.action_deps.echo.print(
-                f"SeCureLI is installed and up-to-date (language = {config.overall_language})"
+                f"SeCureLI is installed and up-to-date (languages = {config.languages})"
             )
             return VerifyResult(
                 outcome=VerifyOutcome.UP_TO_DATE,
@@ -143,9 +142,7 @@ class Action(ABC):
             )
 
         try:
-            metadata = self.action_deps.language_support.apply_support(
-                config.overall_language
-            )
+            metadata = self.action_deps.language_support.apply_support(config.languages)
 
             # Update config with new version installed and save it
             config.version_installed = metadata.version
@@ -203,12 +200,10 @@ class Action(ABC):
 
             # HERE -> decide overall language
             all_languages = analyze_result.language_proportions
-            overall_language = list(analyze_result.language_proportions.keys())[0]
-            self.action_deps.echo.print(
-                f"Overall Detected Language: {overall_language}"
-            )
+            languages = list(analyze_result.language_proportions.keys())
+            self.action_deps.echo.print(f"Overall Detected Language: {languages}")
 
-            metadata = self.action_deps.language_support.apply_support(all_languages)
+            metadata = self.action_deps.language_support.apply_support(languages)
 
         except (ValueError, LanguageNotSupportedError, InstallFailedError) as e:
             self.action_deps.echo.error(
@@ -219,25 +214,25 @@ class Action(ABC):
             )
 
         config = SecureliConfig(
-            overall_language=overall_language,
+            languages=languages,
             version_installed=metadata.version,
         )
         self.action_deps.secureli_config.save(config)
 
         if secret_test_id := metadata.security_hook_id:
             self.action_deps.echo.print(
-                f"{config.overall_language} supports secrets detection; running {secret_test_id}."
+                f"{config.languages} supports secrets detection; running {secret_test_id}."
             )
             self.action_deps.scanner.scan_repo(
                 ScanMode.ALL_FILES, specific_test=secret_test_id
             )
         else:
             self.action_deps.echo.warning(
-                f"{config.overall_language} does not support secrets detection, skipping"
+                f"{config.languages} does not support secrets detection, skipping"
             )
 
         self.action_deps.echo.print(
-            f"SeCureLI has been installed successfully (language = {config.overall_language})"
+            f"SeCureLI has been installed successfully (language = {config.languages})"
         )
 
         return VerifyResult(
